app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt==5.0.0
httpx
python-dotenv
cloudinary
python-multipart
